refactor: Applications can be events or commands or both so it should be called applications instead of commands
